Refactor : User module segmented to Auth Module & User Module
