+++ src/auth/tokenPayload.interface.ts
@@ -1,0 +1,3 @@
+export interface TokenPayload {
+  userId: string;
+}
